finished grid system

+++ src/common/constants.rs
@@ -1,2 +1,2 @@
-pub const ARENA_WIDTH: u8 = 15;
-pub const ARENA_HEIGHT: u8 = 13;
+pub const ARENA_WIDTH: u32 = 15;
+pub const ARENA_HEIGHT: u32 = 13;
